ft_strlen

+++ libft.h
@@ -1,6 +1,8 @@
+#include <stdio.h>
+
 int	ft_isalpha(int c);
 int	ft_isdigit(int c);
 int	ft_isalnum(int c);
 int	ft_isascii(int c);
 int	ft_isprint(int c);
-int	ft_strlen(char *str);
+size_t	ft_strlen(char *str);
